Pint Fixes

+++ src/Http/Controllers/LockRequestController.php
@@ -13,34 +13,30 @@ class LockRequestController extends BaseController
     {
         try {
             $modelLock = ModelLock::findOrFail($lockRequestID);
-            if ($modelLock->user_id == Auth::id())
-            {
-                $modelLock->delete();
+            if (!empty($modelLock) && isset($modelLock->user_id)) {
+                if ($modelLock->user_id == Auth::id()) {
+                    $modelLock->delete();
+                }
             }
         } catch (\Illuminate\Database\Eloquent\ModelNotFoundException $e) {
             session()->flash('message', __('error.modelnotfound'));
             return false;
         }
-
     }
 
     public function request($lockRequestID)
     {
         try {
             $modelLock = ModelLock::findOrFail($lockRequestID);
-            if (!empty(Auth::user()))
-            {
-                if (!empty($modelLock))
-                {
+            if (!empty(Auth::user())) {
+                if (!empty($modelLock)) {
                     $mlw = new ModelLockWatcher();
                     $mlw->model_lock_id = $modelLock->id;
                     $mlw->user_id = Auth::id();
                     $mlw->user_type = get_class(Auth::user());
                     $mlw->save();
                 }
-
             }
-
         } catch (\Illuminate\Database\Eloquent\ModelNotFoundException $e) {
             session()->flash('message', __('error.modelnotfound'));
             return false;
